feat: add connection icons to subscription listings (#3769)
For easier identification of connection issues, we've decided to add an
icon to the subscription listings. The connection is healthy if there is
a connection time and no disconnection time. In all other cases the
connection is unhealthy.

---

For the icons we already have a `name=connected` icon, which is the
`PlugIcon`. The `VitalsIcon` was already used for `health` in one case
and I've decided to reuse this one for `healthy` vs `unhealthy`
(displayed as `RemoveIcon`).

Closes #3183

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>