Revert "chore(ci): reduce the majority of fake-api to a createFetch function" (#3788)
Reverts kumahq/kuma-gui#3734

This causes issues elsewhere, which I'm not sure if are due to our
GithubActions or not. I don't want to block further work today so I'm
reverting this for now. Will bring it back later.

In the meantime I'm gonna start taking a look at those GitHubActions and
make them a little easier to work with, they are the last ones we have
totally refactored so I may as well bite the bullet and finish up these
last ones.

Signed-off-by: John Cowen <john.cowen@konghq.com>