feat: implement x-search component (#3737)
As part of the improvement for the filtering on lists, collections and
tables I have implemented a new filter bar component `XSearch`, which
deprecates the old `FilterBar` component.

Meeting the following requirements:
- Each `tag:value` is highlighted
- The full input value is still one string, which means that as a user I
can click in a `tag:value` and modify/delete what I need to change
without loosing the highlighted `tag:value` pair
- ~On submit the full input value is parsed and returned via
callback/event-handler. Returns raw and parsed values.~
- ~The form is part of the `*View.vue` templates and not part of the
component. I've added a `name` prop which is passed to the nested
`input`, such that on form submit the value can be extracted using
`FormData` on the form event.~ We will revisit the form handling in a
follow-up. For now we stick to `@change`-handlers.

---

Initially I hoped that it's possible to leverage our `XInput` component,
but unfortunately the way how the underlying `KInput` component is
implemented, it does not really allow having a `cover`. At least I
didn't find an easy way to achieve that. So I decided to write it from
scratch again but use the design tokens.

This is the first part and does not include the dropdown.

Closes #3707

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>