chore(ci): adds resolver to resolve node execs from @kumahq/config (#3818)
TLDR; Adds a "node resolver" to our Makefile scripts

---

We've moved all of our CI/dev utility scripts over to Makefiles,
because:

- Its overly difficult writing scripts inside JSON files (package.json)
- Makefiles are sharable, package.json scripts aren't
- All other benefits of Make

The one downside is, Make does not use node import/modules or node
resolution for its binaries.

This means that if we install something in `@kumahq/config` and run one
of `@kumahq/config`s Make scripts that uses a node binary it isn't
necessarily looked up properly because Make uses its `cwd` and `$PATH`
as a "path".

We've been using `npx` which relies on the executable being installed in
root for a Make call.

So from `/packages/kuma-gui` to find an executable in `/`.

Whereas if we install the executable in `@kumahq/config` there is every
chance that the executable will end up in `/packages/config`, so using
`npx` from within `/packages/kuma-gui` will not find for example
`/packages/config/node_modules/.bin/vue-tsc`.

This PR adds a simple node resolver to lookup any Makefile binaries for
any of the make targets we use. We then use that to call `vue-tsc`.

Depending on how successful this is we can roll it out in a similar
fashion across out Makefile targets and move any executable dependencies
of `@kumahq/config` to where they should be (in `@kumahq/config`)

Signed-off-by: John Cowen <john.cowen@konghq.com>