chore(control-planes): move routing to be a child of control-planes (#3795)
Closes https://github.com/kumahq/kuma-gui/issues/3570

After this goes in I'll probably make some immediate amends ontop of
https://github.com/kumahq/kuma-gui/pull/3789, so ideally this would go
in first.

I probably should have done this PR first tbf. The plan is to put the
notification about "control-plane using a memory store" in the
control-plane module where it belongs. The only reason we haven't done
that already is because we had no where to put it that meant its always
visible, so it ended up going in App.vue.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>