feat(notifications): adds dismissal, GC and resetting to notifications (#3789)
At a general level this PR adds a dismissal button to all our
notifications.

If you dismiss a an alert box, all notifications within that box are
saved to localStorage as dismissed for 7 days at which point it is
"un-dismissed" meaning you will see the notification again if you
haven't resolved what it is warning you about.

localStorage values also be removed i.e. "un-dismissed" if we detect
that you have resolved the reason for the warning, meaning if you then
remove whatever you did to resolve the warning, you will be warned
again.

Further notes:

- I moved around AppView a little so I could put all the notification
display all in one place in AppView (instead of also in
ApplicationShell), this also meant I could remove some CSS that we no
longer need (we are now using XLayout for the same layout)
- ~I moved the memory store notification to the `App.vue` file so that
it is visible all the time, this will mean we will have to duplicate
this. Ideally this notification would be in ControlPlaneDetailView, but
the routing structure is incorrect (see
https://github.com/kumahq/kuma-gui/issues/3570)~ We resolved the linked
issue with further work and now the notification is where we want it in
ControlPlaneDetailView


Closes https://github.com/kumahq/kuma-gui/issues/3605

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>