feat(policy-list-view): use x-search (#3827)
Replaces the current custom implementation of a search field with the
new `XSearch` component and allow filtering for `name`, `namespace` and
`zone` (in case it's supported by the respective policy type). I've
updated the mocks to reflect the filter query.

As discussed in https://github.com/kumahq/kuma-gui/pull/3821, in order
to DRY the parsing of the `search` string and as suggested I've added a
`Resources.search` to `resources/data/Resource` to be shared with other
modules.

Part of #3757

Signed-off-by: schogges <moritz.fleck@konghq.com>