chore: turn x into a native vue plugin (#3800)
Turns `x` into a more traditional Vue plugin.

We had a conversation a little while back about how we might move things
into modules, and one thing that was spoken about was whether we should
split up our modules form the service container.

We had two choices each with Pros and Cons:

### 1. Keep the container in the module

Pros: Easier to manage each module. Its only us using the modules and we
always use the container, so pragmatically no massive point in splitting
Cons: modules now need to depend on the container and the utility
functions in order to provide a container.

### 2. Have two modules, a "container module" that depends on its
"vanilla module"

Pros: We have a "vanilla" module that doesn't need to depend on the
container
Cons: 
- We now have 2 modules to manage and support (multiple this by the
amount of modules we will need)
- This is almost like a deeper inheritance chain. Whilst the vanilla
module is dependency free, we still have to depend on the container
which then depends on the vanilla module, just now we have a deeper
dependency tree.

I thought of a third option which is contained (sorry!! 😆 ) in this PR

### 3. Make our container install native Vue plugins and make their
components injectable

Pros: 
- We have a "vanilla" module that doesn't need to depend on the
container
- Easier to manage each module, we have one single module, not a
"container only module" depending on the "vanilla module"
- We could theoretically install any Vue module and have the ability to
inject components.
- Theoretically anyone could use `x` as we do with or without our
container/component injecting.

Note: This only works if you use our container to install the plugin.

---

I had to do a little bit of moving things around to make things make a
bit more sense, please ask if anything isn't clear.

Signed-off-by: John Cowen <john.cowen@konghq.com>