refactor(dataplane-detail-view): move tls card content to about card (#3776)
Moves the content of the TLS card in the `DataplaneDetailView` to the
About Section of the dataplane. I think this is the last of the cards
that use the old style of small title and big value to display data.
I've aligned the look with the rest of the cards and used badges for the
values.

On top of the visual changes, I've added a warning notification when the
TLS certificate is about to expire (1 week).

Before

![image](https://github.com/user-attachments/assets/e658eaa1-0764-43d4-b568-73c6cd45941e)


After

![image](https://github.com/user-attachments/assets/bed7ce9f-d603-428b-baa2-ef7fe0fba767)


Closes #3411

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>