refactor(x-search): use useResizeObserver (#3856)
We hit a flake a couple of times `ResizeObserver loop completed with
undelivered notifications` which could be the result of not cleaning up
the resize observer we register in `XSearch.onMounted`. While thinking
of creating a custom hook to prevent us running into this case again, we
decided to leverage one of `@vueuse`'s exported hooks
`useResizeObserver` which should already include everything we need
instead of reinventing the wheel.

Signed-off-by: schogges <moritz.fleck@konghq.com>