fix(ci): remove dedupe from sync target (#3889)
`npm dedupe` can sometimes break overtime, we added it to our CI so we
could notice when it breaks and therefore investigate. 

Shorter term `npm dedupe` is currently blocking us, so I want to remove
for now, and potentially add back later if we ever need it. If we do
that we will need to do a full `npm upgrade` of all our deps first.

Signed-off-by: John Cowen <john.cowen@konghq.com>