feat(dpp/overview): add info about transparent proxying  (#3798)
In order to inform the user about the potential improvement to configure
`transparentProxying` for the dataplane networking, this PR adds an info
about this in case there is no
`'feature-transparent-proxy-in-dataplane-metadata'` or
`'bind-outbounds'` in `dataplaneInsight.metadata.features` list. The
variant is just an `info` because it could be intentional to not
configure this.
Some notes:
- I've started to integrate the new dataplane types from OAS and I hit
some issues while I tried that. Some of the types are not matching the
actual response, such as `dataplaneInsight.subscriptions[].connectTime`.
- Even though `networking.transparentProxying` is deprecated and will be
removed soon, some users have this configured and therefore I think we
should still include that in the condition of wether to show the info or
not. So a simple check on `'transparentProxying' in
dataplane.networking` should be enough and no need to add this to types
or data-layer.

Closes #2472

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>
Co-authored-by: Charly Molter <charly@molter.io>