feat(dataplane-summary): show inbounds and outbounds in structured (#3896)
Brings back `Inbounds` and `Outbounds` section in summary structured
view.
After removing `inbounds` in #3267 and having it only shown as part of
the config in yaml, we introduced a select to switch the view
(`structured`, `universal`, `kubernetes`). Now that it's up to the user
to decide which type of view is preferred, the `structured` view should
show more info again.

I've partially integrated the dataplane types from OAS.

Closes #2833

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>