refactor(x-search): improve UX (#3902)
Changes the regular expression that highlights the different search
values (filters) to be more user friendly.
It now allows to have a space after a `key:` => `foo: bar` and also
typing `=` instead of `:` is allowed. Still a `=` is being converted
back to `:` behind the scenes.

Generally this aligns more with the current behaviour of dataplane
search. In another follow-up I'll remove the implementation in
dataplanes and connect it to `Resource.search`.

In order to keep track of the regex in just one place, I've created a
mandatory prop for `XSearch`: `searchRegex`.
It's being tracked next to and used in `Resource.search`.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>