refactor(dataplanes): use resource search (#3948)
The dataplanes listings were the last ones not using `Resource.search`.
This is now changing and I've removed the old
`DataplaneOverview.search`. While doing that I had to make sure that we
still keep supporting filter by `tag`, i.e. in some places we still use
tags from `dataplaneTags`. I still wanted to migrate as much as possible
over to use `label`, which I did by using `zone:<zone>` for the
dataplanes listing in `MeshDetailView`.

I've also noticed that in some places I checked for `can('use zones')`
in the `:keys` for `XSearch` and in some places I didn't. This is also
being fixed here.

Also I made sure that in the detail views where we have filter presets
for the dpp listing, that these are removed from the `:keys` of
`XSearch` to prevent users from trying to filter for different values.

Lastly I've fixed a bug we found recently which caused that a
`:value`-prop change of `XSearch` wasn't reflected in the value of the
nested input field.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>