feat(x-search): add validation (#3847)
In order to provide feedback about invalid filters I've added validation
to the `XSearch` component.

- validation happens on a per route/use basis as not all routes support
the same filters.
- a regex for validation is being exported from `Resource` as a base for
common validation
- for the moment the only invalid filters would be a wrong pattern. We
always expect a filter to be `<value>` or `<key>:<value>`, therefore
only `:<value>`, `<key>:` and single `:` are invalid.
- invalid filters are still being added to the query string and sent to
the API. Validation should mainly help the user to understand why
potentially unexpected results are being returned
- The validation is currently only implemented for `DataplanesListView`
and will be rolled out in follow-ups to the other listings.

Closes #3787

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>