feat: add status badge to titles (#3954)
A while back we already had some discussions (x-ref:
https://github.com/kumahq/kuma-gui/pull/3209#issuecomment-2488343300) to
put important information in a more prominent and "persistent" place.
Persistent means that the information would still be shown while
navigating through sub-routes (tabs) of a resource/entity.

This adds the `state` (sometimes also called `status`) to the following
routes:

- ZoneDetails
- ZoneIngressDetails and Summary
- ZoneEgressDetails and Summary
- ServiceDetails
- MeshServiceDetails and Summary
- DelegatedGatewayDetails
- DataplaneDetails and Summary
- ConnectionInboundSummary

---

There is still the component `StatusBadge` in the `common` folder, that
we use in listings and about sections. As we want deprecate the
components and decide wether to move/recreate in `x`, I've not used it
again. I decided to only show the badge. The values of `status`/`state`
of the different resources differs (i.e. `Available` vs `online` ) which
makes it harder to cover all cases in one component and compared to the
amount of templating I thought it's better to just copy&pasta most of
it.
In a follow-up we can remove the usage of `StatusBadge` entirely and
decide again for a new component in `x`.

Closes #3242

Signed-off-by: schogges <moritz.fleck@konghq.com>