chore(mesh): moves mesh/mesh-insights to use OpenAPI (#3968)
We've been a while with a mix of handwritten TS types and OpenAPI
generated ones due to old legacy ProtoBuf being used upstream.

We recently got almost all of the outstanding OpenAPI specs added so we
could hopefully get 100% onto OpenAPI and begin to undo some of our own
old legacy code (HTTP clients and suchlike), but I found that even the
newer specs we've had added are either incorrect or incomplete.

This PR adds a OpenAPI overlay to the upstream OpenAPI specs that we can
use to fix up and otherwise amend the upstream specs. This is better
because we are amending the types using the original format (OpenAPI
YAML not TS which we currently use) which means we can potentially
upstream our amends.

It also means I can just take on what needs to be done to get us to
where we want to be.

Additionally (using the overlays):

- Fixes `Mesh.meshServices.mode` to only be a certain set of strings
- Adds `MeshInsights` to the OpenAPI specs (but does switch to use our
OpenAPI client _yet_)

Signed-off-by: John Cowen <john.cowen@konghq.com>