fix(ci): adds an ignore for ResizeObserver errors in Cypress (#3989)
Sounds like this is a common thing with Cypress, I guess it can happen
also real-life. We have seen it in real-life and added a fix and not
seen it in real-life since then, but we keep getting flakes with it
quite often.

Not super keen on this PR myself, so this might be a temporary thing,
but if we want to remove this ignore we should find a way to stop it
causing flakes first.

Signed-off-by: John Cowen <john.cowen@konghq.com>