chore(renovate): add post-upgrade-task for msw (#3993)
In #3936 `msw` was updated without updating the worker script. I did
this manually in #3988, but ideally this should happen automatically.
Updating the worker script is part of `msw`'s post-install script, but
we use `--ignore-scripts` flag in our make target for installation and I
also noticed that it wouldn't trigger if we run `npm install` from the
repository root. From my point of view this leaves one option, which is
to add a `postUpgradeTasks` for `msw`, which I did here.

Note that in the documentation `postUpgradeTasks` is missing in
`packageRules`, but I ran the config validator and it didn't complain.
So 🤞 that this works and I want to try this with the next bump of `msw`.

Signed-off-by: schogges <moritz.fleck@konghq.com>