feat(zone-detail): add env-icon to title (#3987)
In order to make it more clear to which environment the zone is deployed
to, we want to show the icon of the environment in next to the name of
the zone (title), which is a very prominent spot. In the zone list we
already show this icon in the first column.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>