fix(zone-detail): preferred disconnected subscription config (#4024)
I faced a new flake
https://github.com/kumahq/kuma-gui/actions/runs/15676714259/job/44158981105?pr=4019
which got introduced most likely with
https://github.com/kumahq/kuma-gui/pull/3987.

```sh
AssertionError: Timed out retrying after 4000ms: Expected to find content: 'Universal' within the element: <div.route-view> but never did.
```

We noticed that the data-layer did not consider that subscriptions can
be disconnected and we always use the first subscription in the list to
extract the config. We should rather prefer the `connectedSubscription`
(has no `disconnectTime`) and only fall back to the first subscription
that has a config in case there is no `connectedSubscription`.

---------

Signed-off-by: Mo <moritz.fleck@konghq.com>