chore: simplifies data-source (#3891)
Removes `source()` polling function entirely and replaces it with native
async generators, which is what I originally wanted but didn't get
around to doing.

Polling datasources are now as simple as using async generators
directly:

```typescript
    '/mesh-insights': async function* (params, { signal }) { // AbortController.signal
      while (!signal.aborted) {
         yield MeshInsight.fromCollection(await api.getAllMeshInsights({ size, offset, ...search }))
         await new Promise((resolve) => setTimeout(resolve, 2000))
      }
    },
```

less custom code 🎉

The AbortController could also be passed to the HTTP (or other) client
being used to additionally cancel any outgoing requests when the
DataSource is unmounted.

This also means that our `sources` files only depends on `defineSources`
(for used for typing), and then whatever client you want to use for
actually sourcing the data. In order to do polling or other similar
asynchronous streaming type of interactions you no longer need any
utilities specific to what will soon be our `@kumahq/data` package. Its
all just native vanilla javascript.

---

I also added an `?cacheControl=immutable` here, which is a "one time
request" i.e. during the lifetime of the application the associated
request will only ever happen once, any other requests from a DataSource
will source the data straight from the cache. We don't use this as yet
(and we might never use it), but this usecase wasn't possible with the
previous code so I half used it as a test case.

Lastly, I've added code for exponential retries at a global level (all
datasources) but not enabled it as yet. We'd need to decide when/if and
for what types of errors/error statuses we should use this for.


---

One last note on this I was gonna add some utility functions like this:

```typescript
   const timeout = async (ms: number) => {
        let i: ReturnType<typeof setTimeout>
        await Promise.race([
          new Promise(resolve => i = setTimeout(resolve, ms)),
          new Promise((_, reject) => {
            const cb = () => {
              clearTimeout(i); reject()
            }
            signal.addEventListener('abort', cb, { once: true })
          }),
        ])
      }
...
...
    '/mesh-insights': async function* (params, { signal }, { timeout }) { // <--- here is timeout()
      const { size } = params
      const offset = params.size * (params.page - 1)
      const search = MeshInsight.search(params.search)
      
      while (!signal.aborted) {
        yield MeshInsight.fromCollection(await api.getAllMeshInsights({ size, offset, ...search }))
        await timeout(2000)
      }
    },
```

I'm just very very inclined to keep these arguments very very
simple/native, but I'm very on the fence as something like a timeout
function could be handy (its more correct, but I don't think essential)

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>