chore: uses DataLoader::variant, removes #loadable (#4036)
Closes https://github.com/kumahq/kuma-gui/issues/3259 (also see
https://github.com/kumahq/kuma-gui/pull/3233 mentioned in that issue)

Every removal of `loadable` should be accompanied by an addition of
`variant`, if you don't use a variant="list" on a list type of type, the
only things that happens is you see a spinner rather than a list
skeleton. Despite this being inconsequential, I added/uncommented the
dev-time-only error to nudge you if you ever add an AppCollection that
isn't wrapped in a variant="list".

Also removes the majority of `:items="data?.items ?? [undefined]"`
weirdness mentioned in https://github.com/kumahq/kuma-gui/pull/3233 .
There are a couple of these left that I'll look to remove at some other
point.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>