chore: adds more information when we get a client HTTP error (#4038)
There are a few things I would like to know when we get an unknown
status type error from our HTTP client, are couple are:

1. What was the URL for the request?
2. Is the user still on our page for 3 seconds after the error occurred?

Number one is out of curiosity, and to verify exactly where the request
was going.
Number two is to investigate whether its worth us trying to retry the
request

I have a suspicion some of these errors happen because the user
navigates to another page (via a traditional link, not a JS `pushState`
i.e. a page refresh) whilst a request is in-flight. In this case it
would be pointless retrying. Whereas if we can see via this
console.error that the user is still on the page, it would be worth
retrying.

Signed-off-by: John Cowen <john.cowen@konghq.com>