refactor: move cypress step-definitions to config (#4023)
Moves the cypress step-definitions over to `@kumahq/config`. I've
decided to put it into the config package as we already have cypress
configuration in there and it's just one more file. Generally I think we
should move configs and toolings (as documented) into this package and
once we see that it grows too much, we can split it further down again.

---------

Signed-off-by: Mo <moritz.fleck@konghq.com>