chore(deps): reverts npm upgrade, fixes/exits on npm check (#4080)
We've decided we want to use the version of npm that ships with node,
whatever that is.

This reverts npm back to that version.

I also fixed our `npm` version check, the intention here was have a hard
error when the local npm version does not match the version in the
workspace package.

I have another PR upcoming to remove npm from renovate.

Signed-off-by: John Cowen <john.cowen@konghq.com>