chore(deps): ignore npm from the bot, use the version shipped with node (#4081)
Adds the npm ignore to the top of the package rules, removes the
Node/Npm group that I assume we don't need anymore (?). I think this is
one of those PRs that we won't know is 100% fine until merged.

note for future: I would like to move this file to `.json5` so we can
add comments in it

Signed-off-by: John Cowen <john.cowen@konghq.com>