fix(dataplane/stats): consider inbound servicePort to be different than port (#4066)
When not using transparent proxy it's possible to set a different
`servicePort` than `port`. If not set the `servicePort` internally falls
back to `port`. Due to not considering the potential difference of
`servicePort` we didn't show the stats because the stats refer to the
`servicePort`.
In order to include the potential `servicePort` I've added another field
to the inbounds in the data layer: `clusterName`, which falls back to
the `name` (`localhost_<port>`) which should provide a name of the
grouping between proxy and service based on the networking
configuration.

Closes #4025

---------

Signed-off-by: Mo <moritz.fleck@konghq.com>
Signed-off-by: schogges <moritz.fleck@konghq.com>