chore(deps): removes @types/glob (#4084)
Instead of upgrading this here
https://github.com/kumahq/kuma-gui/pull/4075, I looked to see why we
need this, and I found that the `glob` package is in typescript (at
least version 11 it seems), so I'm pretty sure we don't need to depend
on this anymore.

I also checked in my IDE to make sure I was still getting type hints for
glob after uninstalling this and I am.

Signed-off-by: John Cowen <john.cowen@konghq.com>