refactor: move vite config and shared plugins to config (#4069)
Moves the common parts of the vite config to a base config exposed by
the config package. The different configs don't depend on each other
anymore but each extends the base config individually. Also the common
plugin configs are moved to the config package and each vite config
imports the plugin configs. Also as the `vite.config.development` just
imported and reexported the production config, I have removed it.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>