fix(e2e/dpp/clusters): flake extracting cluster stats of inbound port (#4090)
In https://github.com/kumahq/kuma-gui/pull/4066 we made sure that in
case of a `networking.inbound[].servicePort` that is different from
`.port` we show the correct stats. We've updated our mocks accordingly
to occasionally also set a random `servicePort`. If this is set the
assertion in the test fails because it expects only a `port` being set.
I've also added another test case that covers the `servicePort` being
set to a different value than `port`.

Signed-off-by: schogges <moritz.fleck@konghq.com>