chore(eslint): properly ignore vendor and artifact folders (#4096)
I've noticed that whenever there is an artifact folder in `kuma-gui`
(`dist`) the linter doesn't properly ignore it. I believe this is due to
the configs added after the config that sets the `ignores`. The configs
added after include certain `files` via glob pattern and may override
the `ignores`. I've therefore moved the `ignores` to its own config and
added it as a last config to override all the previous ones. This is now
also more aligned with the official
[documentation/ignoring-files](https://eslint.org/docs/latest/use/configure/migration-guide#ignoring-files).


>```js
>export default [
>	// ...other config
>	{
>		// Note: there should be no other properties in this object
>		ignores: ["**/temp.js", "config/*"],
>	},
>];
>```

Signed-off-by: schogges <moritz.fleck@konghq.com>